Spotify_Scripts/requirements.txt
boto3==1.26.110
numpy==1.24.2
pandas==2.0.0
python-dotenv==1.0.0
spotdl==4.1.4
spotipy==2.23.0